Reorganized project files to allow script generation
