+++ pyproject.toml
@@ -12,7 +12,7 @@ requires-python = ">=3.11, <3.12"
 dependencies = []
 
 [tool.poetry]
-package-mode = false
+# package-mode = false
 
 [tool.poetry.dependencies]
 python = ">=3.11, <3.12"
@@ -27,4 +27,4 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-audio-separator = "main:main"
+audio-separator = "audio_source_separator.main:main"
